add word count method to category implementation

+++ DocumentsObj-C/DocumentsObj-C/Helpers/NSString+WordCount.h
@@ -16,7 +16,7 @@
 
 @interface NSString (WordCount)
 
--(int)wordCount:(NSString*)word;
+-(NSUInteger)wordCount:(NSString*)word;
 
 
 @end
